moved functions from batch.py to utils/batch_utils.py for more robust reference from local.py; added lineage summary statistics to dbstats.json for colouring tree; report rejected genomes to console; added select input to HTML; added --dry-run flag to prevent batch.py from uploading files to server
