+++ index.html
@@ -13,11 +13,10 @@
 </head>
 
 <body>
-  <div class="bar" style="position: fixed; top: 10px; z-index:10" >
+  <div class="bar" style="position: fixed; top: 10px; left: 10px; z-index:10" >
     <div id="search-bar">
       <input type="search" id="search-input"
                placeholder="e.g., EPI_ISL_434070 or Canada">
-      <span class="or">or</span>
       <input id="start-date" class="dates"
                 placeholder="Start">
       to
@@ -31,7 +30,17 @@
       <button type="button" id="previous_button">Previous</button>
       <button type="button" id="next_button">Next</button>
       <button type="button" id="clear_button">Clear</button>
+
       <span onclick="$('#help-search').dialog('open');" style="cursor: help">&#128304;</span>
+
+      <span class="or"></span>
+      <label>Colour by:</label>
+      <select name="tree-colours" id="select-tree-colours">
+        <option>Region</option>
+        <option>No. samples</option>
+        <option>Collection date</option>
+        <option>Divergence</option>
+      </select>
     </div>
   </div>
 
@@ -51,7 +60,7 @@
     <div class="floattitle">
       Time-scaled tree<span onclick="$('#help-timetree').dialog('open');" style="cursor: help">&#128304;</span>
     </div>
-    <div class="floattitle" id="svg-timetreeaxis" style="position: fixed; top: 58px; z-index: 13"></div>
+    <div class="floattitle" id="svg-timetreeaxis" style="position: fixed; top: 80px; z-index: 13"></div>
     <div id="svg-timetree" style="width: 250px; max-width: 250px;"></div>
   </div>
