+++ local.py
@@ -2,15 +2,13 @@ import argparse
 import os
 import sys
 import json
-import subprocess
-from Bio import Phylo
 from datetime import datetime, date
 from csv import DictReader
 
 import covizu
-from covizu.utils import seq_utils
+from covizu.utils import seq_utils, gisaid_utils
 from covizu.utils.progress_utils import Callback
-from batch import *
+from covizu.utils.batch_utils import *
 
 
 def parse_args():
@@ -21,13 +19,10 @@ def parse_args():
     parser.add_argument("infile", type=str,
                         help="input, path to xz-compressed JSON; if not specified, "
                              "download xz file from GISAID provision feed.")
-
     parser.add_argument("lineages", type=argparse.FileType('r'),
                         help="input, CSV output generated by Pangolin")
-
-    parser.add_argument("-o", "--outfile", type=argparse.FileType('w'),
-                        default='data/clusters.{}.json'.format(datetime.now().isoformat().split('.')[0]),
-                        help="output, dest for JSON beadplot file")
+    parser.add_argument("--outdir", type=str, default='data/',
+                        help="option, path to write output files")
 
     parser.add_argument("--bylineage", type=str, default='data/by_lineage.json',
                         help="path to write JSON of features by lineage")
@@ -63,8 +58,6 @@ def parse_args():
     parser.add_argument('--ft2bin', default='fasttree2',
                         help='path to fasttree2 binary executable')
 
-    parser.add_argument('--outdir', default='data/',
-                        help='directory to write TreeTime output files')
     parser.add_argument('--ttbin', default='treetime',
                         help='path to treetime binary executable')
     parser.add_argument('--clock', type=float, default=8e-4,
@@ -110,27 +103,25 @@ def stream_local(path, lineage_file, minlen=29000, mindate='2019-12-01', callbac
         lineages.update({row['taxon']: row['lineage']})
 
     handle = open(path)
-    short = 0
-    baddate = 0
-    nonhuman = 0
+    rejects = {'short': 0, 'baddate': 0, 'nonhuman': 0}
     for header, seq in seq_utils.iter_fasta(handle):
         if len(seq) < minlen:
-            short += 1
+            rejects['short'] += 1
             continue  # sequence is too short
 
         # hCoV-19/Canada/Qc-L00240569/2020|EPI_ISL_465679|2020-03-27
         label, accn, coldate = header.split('|')
         country = label.split('/')[1]
         if country == '' or country[0].islower():
-            nonhuman += 1
+            rejects['nonhuman'] += 1
             continue
 
         if coldate.count('-') != 2:
-            baddate += 1
+            rejects['baddate'] += 1
             continue  # incomplete collection date
         dt = seq_utils.fromisoformat(coldate)
         if dt < mindate or dt > date.today():
-            baddate += 1
+            rejects['baddate'] += 1
             continue  # reject records with non-sensical collection date
 
         lineage = lineages.get(header, None)
@@ -152,9 +143,8 @@ def stream_local(path, lineage_file, minlen=29000, mindate='2019-12-01', callbac
         yield record
 
     if callback:
-        callback("Rejected {} short genomes\n"
-                 "         {} records with bad dates\n"
-                 "         {} non-human genomes".format(short, baddate, nonhuman))
+        callback("Rejected {short} short genomes\n         {baddate} records with bad "
+                 "dates\n         {nonhuman} non-human genomes".format(**rejects))
 
 
 def process_local(args, callback=None):
@@ -188,23 +178,37 @@ if __name__ == "__main__":
         # export to file to process large lineages with MPI
         json.dump(by_lineage, handle)
 
-    timetree = build_timetree(by_lineage, args, cb.callback)
-
-    # FIXME: this is fragile, fails if user specifies custom output file name
-    head, tail = os.path.split(args.outfile.name)
-    timestamp = tail.split('.')[1]
-    nwk_file = os.path.join(head, 'timetree.{}.nwk'.format(timestamp))
+    # reconstruct time-scaled tree
+    timetree, residuals = build_timetree(by_lineage, args, cb.callback)
+    timestamp = datetime.now().isoformat().split('.')[0]
+    nwk_file = os.path.join(args.outdir, 'timetree.{}.nwk'.format(timestamp))
     with open(nwk_file, 'w') as handle:
         Phylo.write(timetree, file=handle, format='newick')
 
+    # generate beadplots and serialize to file
     result = make_beadplots(by_lineage, args, cb.callback, t0=cb.t0.timestamp())
-    args.outfile.write(json.dumps(result))  # serialize results to JSON
+    outfile = open(os.path.join(args.outdir, 'clusters.{}.json'.format(timestamp)), 'w')
+    outfile.write(json.dumps(result))  # serialize results to JSON
+    outfile.close()
 
     # write data stats
-    dbstat_file = os.path.join(head, 'dbstats.{}.json'.format(timestamp))
+    dbstat_file = os.path.join(args.outdir, 'dbstats.{}.json'.format(timestamp))
     with open(dbstat_file, 'w') as handle:
         nseqs = sum([len(rows) for rows in by_lineage.values()])
-        val = {'lastupdate': timestamp.split('T')[0], 'noseqs': nseqs}
+        val = {
+            'lastupdate': timestamp.split('T')[0],
+            'noseqs': nseqs,
+            'lineages': {}
+        }
+        for lineage, samples in by_lineage.items():
+            ndiffs = [len(x['diffs']) for x in samples]
+            val['lineages'][lineage] = {
+                'nsamples': len(samples),
+                'lastcoldate': max(x['covv_collection_date'] for x in samples),
+                'residual': residuals[lineage],
+                'max_ndiffs': max(ndiffs),
+                'mean_ndiffs': sum(ndiffs)/len(ndiffs)
+            }
         json.dump(val, handle)
 
     cb.callback("All done!")
